Refactor with adding new folders
