--- money-market-shoprite/src/main/kotlin/com/example/moneymarketshoprite/Account.kt
@@ -1,7 +1,0 @@
-package com.example.moneymarketshoprite
-
-interface Account {
-    fun handleDeposit(customerDepositCommand: DepositCommand)
-    fun handleTransfer(customerTransferCommand: TransferCommand)
-    fun handleGenerateReport(): List<TransactionReportResponse>
-}
