+++ money-market-shoprite/src/main/kotlin/com/example/moneymarketshoprite/Transaction.kt
@@ -1,0 +1,4 @@
+package com.example.moneymarketshoprite
+
+class Transaction {
+}
