+++ money-market-shoprite/src/main/kotlin/com/example/moneymarketshoprite/models/TransactionReportResponse.kt
@@ -1,3 +1,3 @@
-package com.example.moneymarketshoprite
+package com.example.moneymarketshoprite.models
 
 class TransactionReportResponse(var amount: Double, var currency: String, var balance: Double)
