+++ money-market-shoprite/src/main/kotlin/com/example/moneymarketshoprite/models/TransferCommand.kt
@@ -1,4 +1,4 @@
-package com.example.moneymarketshoprite
+package com.example.moneymarketshoprite.models
 
 class TransferCommand(val depositAmount: Double, val currency: String, destinationAccountNumber: Int)
