+++ money-market-shoprite/src/main/resources/application.properties
@@ -1,1 +1,8 @@
 spring.application.name=money-market-shoprite
+
+spring.datasource.driver-class-name=org.h2.Driver
+spring.datasource.url=jdbc:h2:file:./data/moneyMarketDb
+spring.datasource.username=name
+spring.datasource.password=password
+spring.sql.init.schema-locations=classpath:schema.sql
+spring.sql.init.mode=always
